feat(翻译功能): ✨ 新增功能助手和谷歌翻译模块 * 拆分原来的谷歌翻译重复代码为模块，方便修改
